rename agent directory --> actor
