+++ lens/actor/inner.py
@@ -3,8 +3,8 @@ from __future__ import absolute_import, division, print_function
 import uuid
 import time
 
-import lens.agent.event as event
-from lens.agent.agent import Agent
+import lens.actor.event as event
+from lens.actor.agent import Agent
 
 
 class Process(object):
