+++ lens/processes/transport_composite.py
@@ -1,6 +1,6 @@
 from __future__ import absolute_import, division, print_function
 
-from lens.agent.inner import Process
+from lens.actor.inner import Process
 
 DEFAULT_COLOR = [color/255 for color in [0, 128, 255]]
